refactor: move const variable to constant file, move helper functions to utils.py
