+++ api_inference_community/routes.py
@@ -5,14 +5,13 @@ import os
 import time
 from typing import Any, Dict
 
+from api_inference_community.constants import IMAGE_OUTPUTS
+from api_inference_community.normalizers import normalize_payload
+from api_inference_community.utils import get_metric
 from api_inference_community.validation import (
     AUDIO,
-    AUDIO_INPUTS,
     IMAGE,
-    IMAGE_INPUTS,
-    IMAGE_OUTPUTS,
     ffmpeg_convert,
-    normalize_payload,
     parse_accept,
 )
 from pydantic import ValidationError
@@ -152,34 +151,6 @@ def call_pipe(pipe: Any, inputs, params: Dict, start: float, accept: str) -> Res
         headers=headers,
         status_code=status_code,
     )
-
-
-def get_metric(inputs, task, pipe):
-    if task in AUDIO_INPUTS:
-        return {"x-compute-audio-length": get_audio_length(inputs, pipe.sampling_rate)}
-    elif task in IMAGE_INPUTS:
-        return {"x-compute-images": 1}
-    else:
-        return {"x-compute-characters": get_input_characters(inputs)}
-
-
-def get_audio_length(inputs, sampling_rate: int) -> float:
-    if isinstance(inputs, dict):
-        # Should only apply for internal AsrLive
-        length_in_s = inputs["raw"].shape[0] / inputs["sampling_rate"]
-    else:
-        length_in_s = inputs.shape[0] / sampling_rate
-    return length_in_s
-
-
-def get_input_characters(inputs) -> int:
-    if isinstance(inputs, str):
-        return len(inputs)
-    elif isinstance(inputs, (tuple, list)):
-        return sum(get_input_characters(input_) for input_ in inputs)
-    elif isinstance(inputs, dict):
-        return sum(get_input_characters(input_) for input_ in inputs.values())
-    return 0
 
 
 async def status_ok(request):
