+++ api_inference_community/validation.py
@@ -1,12 +1,9 @@
-import json
-import os
 import subprocess
-from base64 import b64decode
-from io import BytesIO
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Union
 
 import annotated_types
 import numpy as np
+from api_inference_community.constants import INPUTS_MAPPING, PARAMS_MAPPING
 from pydantic import BaseModel, RootModel, Strict, field_validator
 from typing_extensions import Annotated
 
@@ -125,40 +122,6 @@ class StringInput(RootModel):
     root: str
 
 
-PARAMS_MAPPING = {
-    "conversational": SharedGenerationParams,
-    "fill-mask": FillMaskParamsCheck,
-    "text2text-generation": TextGenerationParamsCheck,
-    "text-generation": TextGenerationParamsCheck,
-    "summarization": SummarizationParamsCheck,
-    "zero-shot-classification": ZeroShotParamsCheck,
-}
-
-
-INPUTS_MAPPING = {
-    "conversational": ConversationalInputsCheck,
-    "question-answering": QuestionInputsCheck,
-    "feature-extraction": StringOrStringBatchInputCheck,
-    "sentence-similarity": SentenceSimilarityInputsCheck,
-    "table-question-answering": TableQuestionAnsweringInputsCheck,
-    "tabular-classification": TabularDataInputsCheck,
-    "tabular-regression": TabularDataInputsCheck,
-    "fill-mask": StringInput,
-    "summarization": StringInput,
-    "text2text-generation": StringInput,
-    "text-generation": StringInput,
-    "text-classification": StringInput,
-    "token-classification": StringInput,
-    "translation": StringInput,
-    "zero-shot-classification": StringInput,
-    "text-to-speech": StringInput,
-    "text-to-image": StringInput,
-}
-
-
-BATCH_ENABLED_PIPELINES = ["feature-extraction"]
-
-
 def check_params(params, tag):
     if tag in PARAMS_MAPPING:
         PARAMS_MAPPING[tag].model_validate(params)
@@ -173,72 +136,6 @@ def check_inputs(inputs, tag):
         raise ValueError(f"{tag} is not a valid pipeline.")
 
 
-AUDIO_INPUTS = {
-    "automatic-speech-recognition",
-    "audio-to-audio",
-    "speech-segmentation",
-    "audio-classification",
-}
-AUDIO_OUTPUTS = {
-    "audio-to-audio",
-    "text-to-speech",
-}
-
-
-IMAGE_INPUTS = {
-    "image-classification",
-    "image-segmentation",
-    "image-to-text",
-    "image-to-image",
-    "object-detection",
-    "zero-shot-image-classification",
-}
-IMAGE_OUTPUTS = {
-    "image-to-image",
-    "text-to-image",
-}
-
-
-TEXT_INPUTS = {
-    "conversational",
-    "feature-extraction",
-    "question-answering",
-    "sentence-similarity",
-    "fill-mask",
-    "table-question-answering",
-    "tabular-classification",
-    "tabular-regression",
-    "summarization",
-    "text-generation",
-    "text2text-generation",
-    "text-classification",
-    "text-to-image",
-    "text-to-speech",
-    "token-classification",
-    "zero-shot-classification",
-}
-
-
-AUDIO = [
-    "flac",
-    "ogg",
-    "mp3",
-    "wav",
-    "m4a",
-    "aac",
-    "webm",
-]
-
-
-IMAGE = [
-    "jpeg",
-    "png",
-    "webp",
-    "tiff",
-    "bmp",
-]
-
-
 def parse_accept(accept: str, accepted: List[str]) -> str:
     for mimetype in accept.split(","):
         # remove quality
@@ -250,25 +147,6 @@ def parse_accept(accept: str, accepted: List[str]) -> str:
         if extension in accepted:
             return extension
     return accepted[0]
-
-
-def normalize_payload(
-    bpayload: bytes, task: str, sampling_rate: Optional[int]
-) -> Tuple[Any, Dict]:
-    if task in AUDIO_INPUTS:
-        if sampling_rate is None:
-            raise EnvironmentError(
-                "We cannot normalize audio file if we don't know the sampling rate"
-            )
-        return normalize_payload_audio(bpayload, sampling_rate)
-    elif task in IMAGE_INPUTS:
-        return normalize_payload_image(bpayload)
-    elif task in TEXT_INPUTS:
-        return normalize_payload_nlp(bpayload, task)
-    else:
-        raise EnvironmentError(
-            f"The task `{task}` is not recognized by api-inference-community"
-        )
 
 
 def ffmpeg_convert(
@@ -343,66 +221,3 @@ def ffmpeg_read(bpayload: bytes, sampling_rate: int) -> np.array:
     if audio.shape[0] == 0:
         raise ValueError("Malformed soundfile")
     return audio
-
-
-def normalize_payload_image(bpayload: bytes) -> Tuple[Any, Dict]:
-    from PIL import Image
-
-    try:
-        # We accept both binary image with mimetype
-        # and {"inputs": base64encodedimage}
-        data = json.loads(bpayload)
-        image = data["image"] if "image" in data else data["inputs"]
-        image_bytes = b64decode(image)
-        img = Image.open(BytesIO(image_bytes))
-        return img, data.get("parameters", {})
-    except Exception:
-        pass
-
-    img = Image.open(BytesIO(bpayload))
-    return img, {}
-
-
-DATA_PREFIX = os.getenv("HF_TRANSFORMERS_CACHE", "")
-
-
-def normalize_payload_audio(bpayload: bytes, sampling_rate: int) -> Tuple[Any, Dict]:
-    if os.path.isfile(bpayload) and bpayload.startswith(DATA_PREFIX.encode("utf-8")):
-        # XXX:
-        # This is necessary for batch jobs where the datasets can contain
-        # filenames instead of the raw data.
-        # We attempt to sanitize this roughly, by checking it lives on the data
-        # path (hardcoded in the deployment and in all the dockerfiles)
-        # We also attempt to prevent opening files that are not obviously
-        # audio files, to prevent opening stuff like model weights.
-        filename, ext = os.path.splitext(bpayload)
-        if ext.decode("utf-8")[1:] in AUDIO:
-            with open(bpayload, "rb") as f:
-                bpayload = f.read()
-    inputs = ffmpeg_read(bpayload, sampling_rate)
-    if len(inputs.shape) > 1:
-        # ogg can take dual channel input -> take only first input channel in this case
-        inputs = inputs[:, 0]
-    return inputs, {}
-
-
-def normalize_payload_nlp(bpayload: bytes, task: str) -> Tuple[Any, Dict]:
-    payload = bpayload.decode("utf-8")
-
-    # We used to accept raw strings, we need to maintain backward compatibility
-    try:
-        payload = json.loads(payload)
-        if isinstance(payload, (float, int)):
-            payload = str(payload)
-    except Exception:
-        pass
-
-    parameters: Dict[str, Any] = {}
-    if isinstance(payload, dict) and "inputs" in payload:
-        inputs = payload["inputs"]
-        parameters = payload.get("parameters", {})
-    else:
-        inputs = payload
-    check_params(parameters, task)
-    check_inputs(inputs, task)
-    return inputs, parameters
